+++ src/professions/interfaces/show-template.interface.ts
@@ -14,4 +14,5 @@ export interface ShowTemplate {
   nations: string;
   industries: string[];
   organisations: GroupedTierOneOrganisations;
+  enforcementBodies: string;
 }