+++ src/professions/professions.controller.spec.ts
@@ -6,6 +6,7 @@ import QualificationPresenter from '../qualifications/presenters/qualification.p
 import { createMockI18nService } from '../testutils/create-mock-i18n-service';
 import industryFactory from '../testutils/factories/industry';
 import professionFactory from '../testutils/factories/profession';
+import organisationFactory from '../testutils/factories/organisation';
 import { translationOf } from '../testutils/translation-of';
 import { ProfessionVersionsService } from './profession-versions.service';
 
@@ -14,9 +15,13 @@ import { ProfessionsController } from './professions.controller';
 import { Organisation } from '../organisations/organisation.entity';
 import * as getGroupedTierOneOrganisationsFromProfessionModule from './helpers/get-grouped-tier-one-organisations-from-profession.helper';
 import { GroupedTierOneOrganisations } from './helpers/get-grouped-tier-one-organisations-from-profession.helper';
+import * as getOrganisationsFromProfessionByRoleModule from './helpers/get-organisations-from-profession-by-role';
+
 import { NationsListPresenter } from '../nations/presenters/nations-list.presenter';
 import { Nation } from '../nations/nation';
 import { ShowTemplate } from './interfaces/show-template.interface';
+import { OrganisationRole } from './profession-to-organisation.entity';
+import { organisationList } from './presenters/organisation-list';
 
 jest.mock('../organisations/organisation.entity');
 jest.mock('../nations/presenters/nations-list.presenter');
@@ -74,6 +79,20 @@ describe('ProfessionsController', () => {
         )
         .mockReturnValue(expectedOrganisations);
 
+      const expectedAwardingBodies = organisationFactory.buildList(5);
+      const expectedEnforcementBodies = organisationFactory.buildList(3);
+
+      const getOrganisationsFromProfessionByRoleSpy = jest
+        .spyOn(
+          getOrganisationsFromProfessionByRoleModule,
+          'getOrganisationsFromProfessionByRole',
+        )
+        .mockImplementation((_profession, role) =>
+          role === OrganisationRole.AwardingBody
+            ? expectedAwardingBodies
+            : expectedEnforcementBodies,
+        );
+
       (Organisation.withLatestLiveVersion as jest.Mock).mockImplementation(
         (organisation) => organisation,
       );
@@ -89,9 +108,11 @@ describe('ProfessionsController', () => {
         qualifications: await new QualificationPresenter(
           profession.qualification,
           createMockI18nService(),
+          expectedAwardingBodies,
         ).summaryList(false, true),
         nations: mockNationsHtml,
         industries: [translationOf('industries.example')],
+        enforcementBodies: organisationList(expectedEnforcementBodies),
         organisations: expectedOrganisations,
       } as ShowTemplate);
 
@@ -105,6 +126,16 @@ describe('ProfessionsController', () => {
       expect(
         getGroupedTierOneOrganisationsFromProfessionSpy,
       ).toHaveBeenCalledWith(profession, 'latestLiveVersion');
+      expect(getOrganisationsFromProfessionByRoleSpy).toHaveBeenCalledWith(
+        profession,
+        OrganisationRole.AwardingBody,
+        'latestLiveVersion',
+      );
+      expect(getOrganisationsFromProfessionByRoleSpy).toHaveBeenCalledWith(
+        profession,
+        OrganisationRole.EnforcementBody,
+        'latestLiveVersion',
+      );
     });
 
     describe('publishing with missing data', () => {
@@ -185,6 +216,7 @@ describe('ProfessionsController', () => {
             nations: mockNationsHtml,
             industries: [translationOf('industries.example')],
             organisations: expectedOrganisations,
+            enforcementBodies: organisationList([]),
           } as ShowTemplate);
         });
       });