+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Placeholder dashboard for viewing recognition decision data
 - Placeholder page for showing a single decision dataset
+- Show Awarding Bodies against profession
+- Show Enforcement Bodies against profession
 
 ## [release-015] - 2022-04-06